task3 RobotsGame 4.5

+++ src/main/java/RobotsGame/Target.java
@@ -3,7 +3,7 @@ package RobotsGame;
 import java.awt.Color;
 
 public class Target extends BaseRobotsGameObject {
-    public Target(double x, double y, Color color, Properties properties, int energy)
+    public Target(int x, int y, Color color, Properties properties, int energy)
     {
         this.x = x;
         this.y = y;
